+++ packages/cli/generation/ir-migrations/src/IntermediateRepresentationMigrator.ts
@@ -41,6 +41,7 @@ import { V41_TO_V40_MIGRATION } from "./migrations/v41-to-v40/migrateFromV41ToV4
 import { V42_TO_V41_MIGRATION } from "./migrations/v42-to-v41/migrateFromV42ToV41";
 import { V43_TO_V42_MIGRATION } from "./migrations/v43-to-v42/migrateFromV43ToV42";
 import { V44_TO_V43_MIGRATION } from "./migrations/v44-to-v43/migrateFromV44ToV43";
+import { V45_TO_V44_MIGRATION } from "./migrations/v45-to-v44/migrateFromV45ToV44";
 import { V5_TO_V4_MIGRATION } from "./migrations/v5-to-v4/migrateFromV5ToV4";
 import { V6_TO_V5_MIGRATION } from "./migrations/v6-to-v5/migrateFromV6ToV5";
 import { V7_TO_V6_MIGRATION } from "./migrations/v7-to-v6/migrateFromV7ToV6";
@@ -269,6 +270,7 @@ const IntermediateRepresentationMigrator = {
 
 const INTERMEDIATE_REPRESENTATION_MIGRATOR = IntermediateRepresentationMigrator.Builder
     // put new migrations here
+    .withMigration(V45_TO_V44_MIGRATION)
     .withMigration(V44_TO_V43_MIGRATION)
     .withMigration(V43_TO_V42_MIGRATION)
     .withMigration(V42_TO_V41_MIGRATION)